test(work-adapter): scope FailingWorkItemStore to atomicity test via QuarkusTestProfile (engine#282)
FailingWorkItemStore was registered globally in selected-alternatives,
making it the active WorkItemStore for every @QuarkusTest class in the
module. Move it to HumanTaskScheduleHandlerAtomicityTest with its own
QuarkusTestProfile.getEnabledAlternatives(), where it activates only for
the single test that needs it. HumanTaskScheduleHandlerTest reverts to
the default profile with InMemoryWorkItemStore active.

Closes #282
