+++ work-adapter/src/test/resources/application.properties
@@ -20,5 +20,4 @@ quarkus.arc.selected-alternatives=\
   io.casehub.persistence.memory.MemorySubCaseGroupRepository,\
   io.casehub.persistence.memory.MemoryPlanItemStore,\
   io.casehub.workadapter.NoOpLedgerEntryRepository,\
-  io.casehub.workadapter.NoOpReactiveLedgerEntryRepository,\
-  io.casehub.workadapter.HumanTaskScheduleHandlerTest$FailingWorkItemStore
+  io.casehub.workadapter.NoOpReactiveLedgerEntryRepository
